+++ configuration/config.properties
@@ -18,14 +18,13 @@ environment.hostname.local=http://www.devbox
 #
 # Responsive layout break points
 #
-layout.screensize.small.width=600
-layout.screensize.small.height=800
-layout.screensize.medium.width=700
-layout.screensize.medium.height=800
-layout.screensize.large.width=1100
-layout.screensize.large.height=800
-layout.screensize.xlarge.width=1600
-layout.screensize.xlarge.height=800
+layout.screensize.mobile.width=600
+layout.screensize.mobile.height=800
+layout.screensize.tablet.width=700
+layout.screensize.tablet.height=800
+layout.screensize.desktop.width=1100
+layout.screensize.desktop.height=800
+
 
 
 #
